+++ features/transaction/build.gradle
@@ -27,7 +27,7 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation project(Modules.data_transactions)
+    implementation project(Modules.data_transaction)
     implementation Libraries.kotlin
     implementation Libraries.threeTenAbpJakeWharton
     implementation Libraries.ktx
