+++ features/transaction/proguard-rules.pro
