+++ features/transaction/src/main/res/layout/fragment_transaction.xml
