+++ libraries/data-transaction/src/main/AndroidManifest.xml
