+++ libraries/presentation/build.gradle
@@ -37,7 +37,7 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation project(Modules.data_account)
-    implementation project(Modules.data_transactions)
+    implementation project(Modules.data_transaction)
     implementation Libraries.moshiKotlin
     implementation Libraries.moshiAdapters
     kapt Libraries.moshiCodegen
