+++ libraries/repository/build.gradle
@@ -30,7 +30,7 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation project(Modules.data_account)
-    implementation project(Modules.data_transactions)
+    implementation project(Modules.data_transaction)
 
     implementation Libraries.koinAndroid
     implementation Libraries.koinViewModel
