+++ settings.gradle
@@ -1,1 +1,1 @@
-include Modules.app, Modules.data_account, Modules.data_transactions, Modules.repository, Modules.features_account, Modules.features_transactions, Modules.presentation, Modules.navigation
+include Modules.app, Modules.data_account, Modules.data_transaction, Modules.repository, Modules.features_account, Modules.features_transaction, Modules.presentation, Modules.navigation
